requirements.txt
blinker==1.6.2
cffi==1.16.0
click==8.1.3
colorama==0.4.6
cryptography==42.0.0
exceptiongroup==1.1.1
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
fpdf==1.7.2
greenlet==3.0.3
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pillow==10.2.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pypng==0.20220715.0
pytesseract==0.3.10
pytest==7.3.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.15.0
